+++ src/tmux/mouse.rs
@@ -15,14 +15,14 @@ use super::copy_mode::{CopyModeSnapshot, cell_at_pane, cursor_deltas};
 use super::pane_hit::{cell_at_local, phys_to_pane_local, tmux_pane_at_phys};
 use super::render::{DividerPixelRect, PackedTmuxLayout};
 use crate::configs::OzmuxConfigsResource;
-use crate::inline_webview::{InlineWebview, inline_hit_at, inline_local_dip};
 use crate::input::InputPhase;
 use crate::input::current_modifiers;
 use crate::input::hyperlink::{link_modifier_held, should_open_at, try_open_uri};
-use crate::osc_webview::NonInteractive;
 use crate::picker::SessionPicker;
 use crate::ui::copy_mode::CopyModeState;
 use crate::ui::copy_search::CopyPrompt;
+use crate::webview::inline::{InlineWebview, inline_hit_at, inline_local_dip};
+use crate::webview::osc::NonInteractive;
 use bevy::ecs::system::SystemParam;
 use bevy::input::ButtonState;
 use bevy::input::mouse::MouseButtonInput;
